Refactor chat.js to improve error handling and ensure data consistency during message deletion and score updates
- Moved local state updates to occur only after successful backend deletion in pruneBranch function.
- Added error alerts for failed score updates in createScoreSelect function.
- Enhanced error handling in clearChat function to notify users of deletion failures.
- Modified appendMessageToDB to return a success status and handle errors appropriately.
- Updated sendMessage function to manage message state based on server response.
